move stack trace into utils
