+++ src/main.cpp
@@ -19,21 +19,6 @@ using std::string;
 
 const std::string GAME_FILE_DIR = "/home/seth/Documents/stase_lichess";
 
-void print_stack_trace(int sig) {
-
-    void *array[25];
-    size_t size;
-
-    // get void*'s for all entries on the stack
-    size = backtrace(array, 25);
-
-    // print out all the frames
-    fprintf(stdout, "Error: signal %d:\n", sig);
-    backtrace_symbols_fd(array, size, STDOUT_FILENO);
-
-    exit(1);
-}
-
 const std::string welcome_message =
         "Welcome to Stase v4.0\n";
